requirements.txt
certifi==2020.12.5
cffi==1.14.5
cryptography==3.4.6
pycparser==2.20
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
peewee==3.14.1
PyMySQL==1.0.2
requests==2.25.1
six==1.15.0
urllib3==1.26.3
vk-api==11.9.1
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
pyyaml==5.4.1
